Deleted UseCaseModule.kt, fixed issues
