--- app/src/main/java/com/oracle/visualize/di/UseCaseModule.kt
@@ -1,27 +1,0 @@
-package com.oracle.visualize.di
-
-import com.oracle.visualize.domain.repositories.AuthRepository
-import com.oracle.visualize.domain.usecases.LoginUseCase
-import com.oracle.visualize.domain.usecases.RegisterUseCase
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object UseCaseModule {
-
-    @Provides
-    @Singleton
-    fun provideLoginUseCase(repository: AuthRepository): LoginUseCase {
-        return LoginUseCase(repository)
-    }
-
-    @Provides
-    @Singleton
-    fun provideRegisterUseCase(repository: AuthRepository): RegisterUseCase {
-        return RegisterUseCase(repository)
-    }
-}
